% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_group}
\alias{add_group}
\title{Add group to Pairs file}
\usage{
add_group(pairfile, g = c("mr", "mg", "er", "eg"))
}
\arguments{
\item{pairfile}{File with selected genes pairs}

\item{group}{group to be added}
}
\value{
Overwrite the file adding group
}
\description{
Add group to Pairs file
}
