% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_imp_scores}
\alias{add_imp_scores}
\title{Merge selected gene pairs with their importance scores}
\usage{
add_imp_scores(impfile, pairsfile)
}
\arguments{
\item{impfile}{file with importance scores}

\item{pairsfile}{file with selected gene pairs}
}
\value{
dataframe with selected gene pairs along with their importance scores
}
\description{
Merge selected gene pairs with their importance scores
}
