% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_imp_scores_loop}
\alias{add_imp_scores_loop}
\title{Merge selected gene pairs with their importance scores}
\usage{
add_imp_scores_loop(path)
}
\arguments{
\item{path}{directory containing Selected Pairs files and Importance file}
}
\value{
Overwrite Selected gene pairs with their importance scores
}
\description{
Merge selected gene pairs with their importance scores
}
