% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_interaction_freq}
\alias{compute_interaction_freq}
\title{Calculate fraction of interacting gene pairs over total}
\usage{
compute_interaction_freq(data, interactions)
}
\arguments{
\item{data}{Dataframe with a subset of gene pairs from \link{subset_by_threshold}}

\item{interactions}{Dataframe with StringDB interactions from \link{get_stringdb_interactions}}
}
\value{
Dataframe with fraction of interacting gene pairs over total by group
}
\description{
Calculate fraction of interacting gene pairs over total
}
