% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_essential_genes}
\alias{filter_essential_genes}
\title{Filter essential genes}
\usage{
filter_essential_genes(data, cutoff)
}
\arguments{
\item{data}{dataframe with selected gene pairs}

\item{cutoff}{value to use as a threshold for the dependecy score coefficient of variantion}
}
\value{
Dataframe with essential genes removed
}
\description{
Filter essential genes
}
