% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find_fun.R
\name{find_fun}
\alias{find_fun}
\title{Finds the name of the package that a function is from.}
\usage{
find_fun(fun)
}
\arguments{
\item{fun}{function or character. The function to find the name of the package for.}
}
\description{
Uses Hadley's "where" from pryr, but gets just the name of the package and
can handle passing in a function instead of just a character string.
Also can handle a string that includes the function name (e.g., 'paris::find_fun' or 'paris/find_fun')
}

