% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find_variables.R
\name{find_variables}
\alias{find_variables}
\title{Finds variables within an expression.}
\usage{
find_variables(expr)
}
\arguments{
\item{expr.}{The expression to evaluate.}
}
\description{
Finds variables within an expression.
}

