% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/function_from_expr.R
\name{function_from_expr}
\alias{function_from_expr}
\title{Takes an expression and turns it into a function by inferring parameters.}
\usage{
function_from_expr(expr)
}
\arguments{
\item{expr}{expression. The expression to turn into a function.}
}
\description{
Takes an expression and turns it into a function by inferring parameters.
}

