% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_paralog_biomart}
\alias{get_paralog_biomart}
\title{Retrieve Paralogs from gene vector}
\usage{
get_paralog_biomart(x)
}
\arguments{
\item{x}{vector of gene names}
}
\value{
Dataframe of paralogs in the gene vector
}
\description{
Retrieve Paralogs from gene vector
}
