% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_stringdb_interactions}
\alias{get_stringdb_interactions}
\title{Retrieve StringDB interactions for gene pairs}
\usage{
get_stringdb_interactions(data)
}
\arguments{
\item{dataframe}{Dataframe with a subset of gene pairs from \link{subset_by_threshold}}
}
\value{
Dataframe with StringDB interactions (if any) of the selected gene pairs
}
\description{
Retrieve StringDB interactions for gene pairs
}
