% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{HTBreaks}
\alias{HTBreaks}
\title{Compute and plot HTBreaks}
\usage{
HTBreaks(data, by.group = TRUE)
}
\arguments{
\item{data}{dataframe with selected pairs from \link{select_pairs}}

\item{by.group}{Boolean, if group by}
}
\value{
Compute Head Tail Breaks and plot density distribution and breaks as vertical lines
}
\description{
Compute and plot HTBreaks
}
