% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_imp_scores}
\alias{load_imp_scores}
\title{Load importance scores}
\usage{
load_imp_scores(path, group = c("mr", "mg", "er", "eg"))
}
\arguments{
\item{path}{directory to Boruta models}

\item{group}{group to be added}
}
\value{
Return dataframe importance score gene pairs
}
\description{
Load importance scores
}
