% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_imp_scores2}
\alias{load_imp_scores2}
\title{Load importance scores}
\usage{
load_imp_scores2(path)
}
\arguments{
\item{path}{directory to Boruta models}
}
\value{
Return dataframe importance score gene pairs
}
\description{
Load importance scores
}
