% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myinfo}
\alias{myinfo}
\title{Info data.}
\format{
A data frame with 1736 rows and 23 variables:
\describe{
\item{DepMap_ID}{DepMap_ID}
\item{stripped_cell_line_name}{stripped_cell_line_name}
\item{CCLE Name}{CCLE Name}
\item{alias}{alias}
\item{COSMIC_ID}{COSMIC_ID}
\item{lineage}{lineage}
\item{lineage_subtype}{lineage_subtype}
\item{lineage_sub_subtype}{lineage_sub_subtype}
\item{sex}{sex}
\item{source}{source}
\item{Achilles_n_replicates}{Achilles_n_replicates}
\item{cell_line_NNMD}{cell_line_NNMD}
\item{culture_type}{culture_type}
\item{culture_medium}{culture_medium}
\item{cas9_activity}{cas9_activity}
\item{RRID}{RRID}
\item{sample_collection_site}{sample_collection_site}
\item{primary_or_metastasis}{primary_or_metastasis}
\item{disease}{disease}
\item{disease_subtype}{disease_subtype}
\item{age}{age}
\item{Sanger_model_ID}{Sanger_model_ID}
\item{additional_info}{additional_info}

...
}
}
\source{
\url{https://depmap.org/portal/}
}
\usage{
myinfo
}
\description{
A dataset containing the cell lines info data.
}
\keyword{datasets}
