% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mynewcn}
\alias{mynewcn}
\title{Copy number data.}
\format{
A data frame with 1657 rows and 3 variables:
\describe{
\item{DepMap_ID}{DepMap_ID}
\item{gene}{gene name}
\item{CN}{CN, Copy number}
...
}
}
\source{
\url{https://depmap.org/portal/}
}
\usage{
mynewcn
}
\description{
A dataset containing the gene copy number data.
}
\keyword{datasets}
