% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mynewexp}
\alias{mynewexp}
\title{Expression data.}
\format{
A data frame with 23164240 rows and 3 variables:
\describe{
\item{DepMap_ID}{DepMap_ID}
\item{gene}{gene name}
\item{TPM}{TPM, Transcripts Per Kilobase Million}
...
}
}
\source{
\url{https://depmap.org/portal/}
}
\usage{
mynewexp
}
\description{
A dataset containing the gene expression data.
}
\keyword{datasets}
