% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mynewmerge}
\alias{mynewmerge}
\title{Dependency data.}
\format{
A data frame with 11458125 rows and 3 variables:
\describe{
\item{DepMap_ID}{DepMap_ID}
\item{gene}{gene name}
\item{value}{value, gene dependency}
...
}
}
\source{
\url{https://depmap.org/portal/}
}
\usage{
mynewmerge
}
\description{
A dataset containing the gene dependency data.
}
\keyword{datasets}
