% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mynewmut}
\alias{mynewmut}
\title{Mutation data.}
\format{
A data frame with 716951 rows and 2 variables:
\describe{
\item{DepMap_ID}{DepMap_ID}
\item{gene}{mutated gene name}
...
}
}
\source{
\url{https://depmap.org/portal/}
}
\usage{
mynewmut
}
\description{
A dataset containing the gene mutation data.
}
\keyword{datasets}
