% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mystat}
\alias{mystat}
\title{Stats data.}
\format{
A data frame with 18333 rows and 6 variables:
\describe{
\item{gene}{gene name}
\item{med}{gene dependency median}
\item{cv}{gene dependency, coefficient of variation}
\item{sd}{gene dependency, standard deviation}
\item{cvmod}{gene dependency, coefficient of variation modified}
\item{r}{gene dependency, range}
...
}
}
\source{
\url{https://depmap.org/portal/}
}
\usage{
mystat
}
\description{
A dataset containing the gene dependency statistics data.
}
\keyword{datasets}
