% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.box.def.dep}
\alias{plot.box.def.dep}
\title{Plot boxplot dependecy vs deficient}
\usage{
\method{plot}{box.def.dep}(depgene, defgene)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{defgene}{Expression gene name}
}
\value{
Plot boxplot dependecy score, by deficiency status (mutated or copy number loss)
}
\description{
Plot boxplot dependecy vs deficient
}
