% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.box.mut.dep}
\alias{plot.box.mut.dep}
\title{Plot boxplot dependecy vs mutation}
\usage{
\method{plot}{box.mut.dep}(depgene, mutgene)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{mutgene}{Expression gene name}
}
\value{
Plot boxplot dependecy score, by mutational status plus jitter
}
\description{
Plot boxplot dependecy vs mutation
}
