% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.rank.def.ratio.dep.by.disease}
\alias{plot.rank.def.ratio.dep.by.disease}
\title{Plot ranked barplot dependecy score, colored by ratio expression of two genes and red dots in deficient cells by disease}
\usage{
\method{plot}{rank.def.ratio.dep.by.disease}(depgene, expgenes = c("TYMS", "TYMP"), defgene)
}
\arguments{
\item{depgene}{Gene dependecy score}

\item{expgenes}{Genes ratio expression}

\item{defgene}{Gene deficient (mutated or copy number loss)}
}
\value{
Plot ranked barplot dependecy score, colored by expression and red dots in deficient cells bydisease
}
\description{
Plot ranked barplot dependecy score, colored by ratio expression of two genes and red dots in deficient cells by disease
}
