% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.rank.def.ratio.dep}
\alias{plot.rank.def.ratio.dep}
\title{Plot ranked barplot dependecy score, colored by ratio expression of two genes}
\usage{
\method{plot}{rank.def.ratio.dep}(depgene, expgenes = c("TYMS", "TYMP"))
}
\arguments{
\item{depgene}{Gene dependecy score}

\item{expgenes}{Genes ratio expression}
}
\value{
Plot ranked barplot dependecy score, colored by expression
}
\description{
Plot ranked barplot dependecy score, colored by ratio expression of two genes
}
