% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.rank.exp.dep}
\alias{plot.rank.exp.dep}
\title{Plot rank dependecy vs expression}
\usage{
\method{plot}{rank.exp.dep}(depgene, expgene)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{expgene}{Expression gene name}
}
\value{
Plot ranked barplot dependecy score, colored by expression
}
\description{
Plot rank dependecy vs expression
}
