% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.rank.mut.dep.sub}
\alias{plot.rank.mut.dep.sub}
\title{Plot rank dependecy vs mutation}
\usage{
\method{plot}{rank.mut.dep.sub}(depgene, mutgene, cells)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{mutgene}{Expression gene name}

\item{cells}{Cell lines DepMap to subset}
}
\value{
Plot ranked barplot dependecy score, colored by mutational status
}
\description{
Plot rank dependecy vs mutation
}
