% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.sca.exp.dep.ratio}
\alias{plot.sca.exp.dep.ratio}
\title{Plot scatterplot dependecy vs ratio expression}
\usage{
\method{plot}{sca.exp.dep.ratio}(depgene, expgenes = c("TYMS", "TYMP"))
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{expgenes}{Expression gene name}
}
\value{
Plot scatterplot dependecy score of one gene vs the ratio of expression of two genes
}
\description{
Plot scatterplot dependecy vs ratio expression
}
