% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.sca.exp.dep}
\alias{plot.sca.exp.dep}
\title{Plot scatterplot dependecy vs expression}
\usage{
\method{plot}{sca.exp.dep}(depgene, expgene)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{expgene}{Expression gene name}
}
\value{
Plot scatterplot dependecy score vs expression
}
\description{
Plot scatterplot dependecy vs expression
}
