% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.sca.mut.dep}
\alias{plot.sca.mut.dep}
\title{Plot scatterplot dependecy vs mutation}
\usage{
\method{plot}{sca.mut.dep}(depgene, mutgene)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{mutgene}{Expression gene name}
}
\value{
Plot scatterplot dependecy score vs mutation
}
\description{
Plot scatterplot dependecy vs mutation
}
