% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.stringdb.interactions.by.group}
\alias{plot.stringdb.interactions.by.group}
\title{Plot percentage of interacting gene over total}
\usage{
\method{plot}{stringdb.interactions.by.group}(interaction_freq)
}
\arguments{
\item{interaction_freq}{dataframe from \link{compute_interaction_freq}}
}
\value{
Plot percentage of interacting gene over total by group
}
\description{
Plot percentage of interacting gene over total
}
