% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.stringdb.interactions.combined.score.by.group}
\alias{plot.stringdb.interactions.combined.score.by.group}
\title{Plot StringDB combined scores}
\usage{
\method{plot}{stringdb.interactions.combined.score.by.group}(interactions)
}
\arguments{
\item{interactions}{dataframe with interaction for each gene pairs from \link{get_stringdb_interactions}}
}
\value{
Plot StringDB combined scores of interactiong gene pairs
}
\description{
Plot StringDB combined scores
}
