% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.top.rank}
\alias{plot.top.rank}
\title{Plot top rank selected genes pairs}
\usage{
\method{plot}{top.rank}(data, g = c("mr", "mg", "mc", "er", "eg", "ec"), cutoff)
}
\arguments{
\item{data}{dataframe with selected pairs from \link{select_pairs}}

\item{g}{group to look for}

\item{cutoff}{cutoff of scaled importance score}
}
\value{
Plot top rank selected gene pairs for a group with scaled importance score above the cutoff
}
\description{
Plot top rank selected genes pairs
}
