% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot.vl.mut.dep}
\alias{plot.vl.mut.dep}
\title{Plot violin dependecy vs mutation}
\usage{
\method{plot}{vl.mut.dep}(depgene, mutgene)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{mutgene}{Expression gene name}
}
\value{
Plot violin dependecy score, by mutational status plus jitter
}
\description{
Plot violin dependecy vs mutation
}
