% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_box_def_dep}
\alias{plot_box_def_dep}
\title{Plot boxplot dependecy vs deficient}
\usage{
plot_box_def_dep(depgene, defgene)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{defgene}{Expression gene name}
}
\value{
Plot boxplot dependecy score, by deficiency status (mutated or copy number loss)
}
\description{
Plot boxplot dependecy vs deficient
}
