% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_common_pairs}
\alias{plot_common_pairs}
\title{Plot scatterplot Gini corrected vs raw permutation for a group}
\usage{
plot_common_pairs(data, g = c("e", "m"), imp_gini = c("g", "c"))
}
\arguments{
\item{data}{dataframe with selected pairs from \link{select_pairs}}

\item{g}{group: expression or mutation}

\item{imp_gini}{gini or corrected gini}
}
\value{
Plot scatterplot Gini corrected vs raw permutation for expression or mutation data
}
\description{
Plot scatterplot Gini corrected vs raw permutation for a group
}
