% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_rank_def_exp_dep}
\alias{plot_rank_def_exp_dep}
\title{Plot ranked barplot dependecy score, colored by expression and red dots in deficient cells}
\usage{
plot_rank_def_exp_dep(depgene, expgene, defgene)
}
\arguments{
\item{depgene}{Gene dependecy score}

\item{expgene}{Gene expression}

\item{defgene}{Gene deficient (mutated or copy number loss)}
}
\value{
Plot ranked barplot dependecy score, colored by expression and red dots in deficient cells
}
\description{
Plot ranked barplot dependecy score, colored by expression and red dots in deficient cells
}
