% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_rank_def_ratio_dep_by_disease}
\alias{plot_rank_def_ratio_dep_by_disease}
\title{Plot ranked barplot dependecy score, colored by ratio expression of two genes and red dots in deficient cells by disease}
\usage{
plot_rank_def_ratio_dep_by_disease(
  depgene,
  expgenes = c("TYMS", "TYMP"),
  defgene
)
}
\arguments{
\item{depgene}{Gene dependecy score}

\item{expgenes}{Genes ratio expression}

\item{defgene}{Gene deficient (mutated or copy number loss)}
}
\value{
Plot ranked barplot dependecy score, colored by expression and red dots in deficient cells bydisease
}
\description{
Plot ranked barplot dependecy score, colored by ratio expression of two genes and red dots in deficient cells by disease
}
