% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_rank_exp_dep_sub}
\alias{plot_rank_exp_dep_sub}
\title{Plot rank dependecy vs expression}
\usage{
plot_rank_exp_dep_sub(depgene, expgene, cells)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{expgene}{Expression gene name}

\item{cells}{Cell lines DepMap to subset}
}
\value{
Plot ranked barplot dependecy score, colored by expression, cell lines subset
}
\description{
Plot rank dependecy vs expression
}
