% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_rank_mut_dep_sub}
\alias{plot_rank_mut_dep_sub}
\title{Plot rank dependecy vs mutation}
\usage{
plot_rank_mut_dep_sub(depgene, mutgene, cells)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{mutgene}{Expression gene name}

\item{cells}{Cell lines DepMap to subset}
}
\value{
Plot ranked barplot dependecy score, colored by mutational status
}
\description{
Plot rank dependecy vs mutation
}
