% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_sca_exp_dep}
\alias{plot_sca_exp_dep}
\title{Plot scatterplot dependecy vs expression}
\usage{
plot_sca_exp_dep(depgene, expgene)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{expgene}{Expression gene name}
}
\value{
Plot scatterplot dependecy score vs expression
}
\description{
Plot scatterplot dependecy vs expression
}
