% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_sca_exp_dep_ratio}
\alias{plot_sca_exp_dep_ratio}
\title{Plot scatterplot dependecy vs ratio expression}
\usage{
plot_sca_exp_dep_ratio(depgene, expgenes = c("TYMS", "TYMP"))
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{expgenes}{Expression gene name}
}
\value{
Plot scatterplot dependecy score of one gene vs the ratio of expression of two genes
}
\description{
Plot scatterplot dependecy vs ratio expression
}
