% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_sca_mut_dep}
\alias{plot_sca_mut_dep}
\title{Plot scatterplot dependecy vs mutation}
\usage{
plot_sca_mut_dep(depgene, mutgene)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{mutgene}{Expression gene name}
}
\value{
Plot scatterplot dependecy score vs mutation
}
\description{
Plot scatterplot dependecy vs mutation
}
