% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_stringdb_interactions_by_group}
\alias{plot_stringdb_interactions_by_group}
\title{Plot percentage of interacting gene over total}
\usage{
plot_stringdb_interactions_by_group(interaction_freq)
}
\arguments{
\item{interaction_freq}{dataframe from \link{compute_interaction_freq}}
}
\value{
Plot percentage of interacting gene over total by group
}
\description{
Plot percentage of interacting gene over total
}
