% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_stringdb_interactions_combined_score_by_group}
\alias{plot_stringdb_interactions_combined_score_by_group}
\title{Plot StringDB combined scores}
\usage{
plot_stringdb_interactions_combined_score_by_group(interactions)
}
\arguments{
\item{interactions}{dataframe with interaction for each gene pairs from \link{get_stringdb_interactions}}
}
\value{
Plot StringDB combined scores of interactiong gene pairs
}
\description{
Plot StringDB combined scores
}
