% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_top_rank_color}
\alias{plot_top_rank_color}
\title{Plot top rank selected genes pairs colored}
\usage{
plot_top_rank_color(
  data,
  g = c("mr", "mg", "mc", "er", "eg", "ec"),
  color = c("mr", "mg", "mc", "er", "eg", "ec"),
  cutoff
)
}
\arguments{
\item{data}{dataframe with selected pairs from \link{select_pairs}}

\item{g}{group to look for}

\item{color}{second group}

\item{cutoff}{cutoff of scaled importance score}
}
\value{
Plot top rank selected gene pairs for a group with scaled importance score above the cutoff,
colored by another scaled importance score group
}
\description{
Plot top rank selected genes pairs colored
}
