% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_upset}
\alias{plot_upset}
\title{Upsetplot}
\usage{
plot_upset(data, g = c("e", "m"), cutoff)
}
\arguments{
\item{data}{dataframe with selected pairs from \link{select_pairs}}

\item{g}{group: expression or mutationa}

\item{cutoff}{cutoff of scaled importance score}
}
\value{
Subset the dataframe based on the cutoff and plot groups intersections as upsetplot
}
\description{
Upsetplot
}
