% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{plot_vl_mut_dep}
\alias{plot_vl_mut_dep}
\title{Plot violin dependecy vs mutation}
\usage{
plot_vl_mut_dep(depgene, mutgene)
}
\arguments{
\item{depgene}{Dependecy score gene name}

\item{mutgene}{Expression gene name}
}
\value{
Plot violin dependecy score, by mutational status plus jitter
}
\description{
Plot violin dependecy vs mutation
}
