% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotboruta}
\alias{plotboruta}
\title{plot Boruta top 10 features}
\usage{
plotboruta(
  x,
  colCode = c("green", "yellow", "red", "blue"),
  sort = TRUE,
  whichShadow = c(TRUE, TRUE, TRUE),
  col = NULL,
  xlab = "Attributes",
  ylab = "Importance",
  ...
)
}
\arguments{
\item{x}{Boruta object}
}
\value{
Plot top 10 features of a Boruta object
}
\description{
plot Boruta top 10 features
}
