% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_feature_selection.R
\name{RF_mut}
\alias{RF_mut}
\title{Random Forest Feature Selection - Mutated genes -> Dependent gene}
\usage{
RF_mut(data, mut, imp.score = c("raw", "Gini", "Ginicorrected"), outdir)
}
\arguments{
\item{data}{Dataframe with dependecy scores}

\item{mut}{Dataframe with mutational data}

\item{imp.score}{importance score methods; NB raw/Gini for Boruta; Ginicorrected RF}

\item{outdir}{Directory where to write Models}
}
\value{
Write Boruta/Random Forest Feature selection models and selected gene paris in the outdir
}
\description{
Random Forest Feature Selection - Mutated genes -> Dependent gene
}
