% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{run_PARIS}
\alias{run_PARIS}
\title{Main function to run PARIS pipeline}
\usage{
run_PARIS(
  imp.score = c("raw", "Gini", "Ginicorrected"),
  omic = c("mutation", "expression"),
  outdir,
  genelist_dep,
  genelist_feat
)
}
\arguments{
\item{imp.score}{importance score methods; NB raw/Gini for Boruta; Ginicorrected RF}

\item{omic}{Features data to use: Expression or Mutation}

\item{outdir}{Write Boruta/Random Forest Feature selection models and selected gene paris in the outdir}

\item{genelist_dep}{Gene names vector to use as response variables (dependency scores), ALL if NULL}

\item{genelist_feat}{Gene names vector to use as features (Expression or Mutation), ALL if NULL}
}
\value{
Write Boruta/Random Forest Feature selection models and selected gene paris in the outdir
}
\description{
Main function to run PARIS pipeline
}
