% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale01}
\alias{scale01}
\title{Scaling function}
\usage{
scale01(x)
}
\arguments{
\item{x}{vector of importance scores}
}
\value{
The function return scaled importance scores between 0 and 1
}
\description{
Scaling function
}
