% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_pairs}
\alias{select_pairs}
\title{Read selected gene pairs from a directory}
\usage{
select_pairs(path, coeff.filter = TRUE)
}
\arguments{
\item{path}{directory with selected gene pairs files}

\item{coeff.filter}{Boolean, if to apply the Pearson coefficient-based filtering step}
}
\value{
dataframe with all selected gene pairs
}
\description{
Read selected gene pairs from a directory
}
