% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_by_threshold}
\alias{subset_by_threshold}
\title{Subset a dataframe of gene pairs by threshold}
\usage{
subset_by_threshold(data, g = c("e", "m"), threshold)
}
\arguments{
\item{g}{group: expression or mutation}

\item{threshold}{cutoff for the scaled importance score}

\item{mydata}{Dataframe of selected gene pairs from \link{select_pairs}}
}
\value{
subset dataframe
}
\description{
Subset a dataframe of gene pairs by threshold
}
