% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syntax_tree.R
\name{syntax_tree}
\alias{syntax_tree}
\title{Returns the syntax tree for an expression.}
\usage{
syntax_tree(expr)
}
\arguments{
\item{expr.}{The expression to evaluate.}
}
\description{
Returns the syntax tree for an expression.
}

